+++ src/core/TADImagen.h
@@ -15,13 +15,13 @@
 class Imagen
 {
 private:
-  unsigned int ancho;            ///< Ancho de la imagen
-  unsigned int alto;             ///< Alto de la imagen
-  unsigned short max_intensidad; ///< Maxima intensidad de iluminación en la imagen
+  int ancho;            ///< Ancho de la imagen
+  int alto;             ///< Alto de la imagen
+  int max_intensidad; ///< Maxima intensidad de iluminación en la imagen
   std::string path;              ///< Ubicación de la imagen cargada
 
   // Contenido de la imagen
-  std::vector<std::vector<unsigned short>> contenido; ///< Matriz 2D que contiene la imagen, cada valor esta entre 0 y 255, representa que tan iluminado esta el pixel
+  std::vector<std::vector<int>> contenido; ///< Matriz 2D que contiene la imagen, cada valor esta entre 0 y 255, representa que tan iluminado esta el pixel
 
 public:
   /**
@@ -31,7 +31,7 @@ public:
    * @param r_max_intensidad Maxima intesidad de iluminación en la imagen
    * @param r_contenido Matriz 2D que contiene la imagen, cada valor esta entre 0 y 255, representa que tan iluminado esta el pixel
    */
-  Imagen(std::string r_path, unsigned int r_ancho, unsigned int r_alto, unsigned short r_max_intensidad, std::vector<std::vector<unsigned short>> r_contenido);
+  Imagen(std::string r_path, int r_ancho, int r_alto, int r_max_intensidad, std::vector<std::vector<int>> r_contenido);
 
   /**
    * @brief Retorna el path de la imagen
@@ -43,25 +43,25 @@ public:
    * @brief Retorna el ancho de la imagen
    * @return Ancho de la imagen
    */
-  unsigned int getAncho();
+  int getAncho();
 
   /**
    * @brief Retorna el alto de la imagen
    * @return Alto de la imagen
    */
-  unsigned int getAlto();
+  int getAlto();
 
   /**
    * @brief Retorna la intensidad maxima de la imagen
    * @return Intensidad maxima de la imagen
    */
-  unsigned short getMaxIntensidad();
+  int getMaxIntensidad();
 
   /**
    * @brief Retorna el contenido de la imagen
    * @return Contenido de la imagen
    */
-  std::vector<std::vector<unsigned short>> getContenido();
+  std::vector<std::vector<int>> getContenido();
 };
 
 #endif
